Move some stuff into a more manageable file
